backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
groq==0.5.0
python-dotenv==1.0.1
pandas==2.2.1
